the customIconTabBar component is renamed to customTabBarIcon. styles are extracted to a styles.ts file. active icon files are removed (tintColor prop is used). the icon search is extracted to a utils.ts file
