+++ wbooks/src/app/assets/tabBar/index.ts
@@ -1,25 +1,15 @@
 import icAddNew from './icAddNew.png';
-import icAddNewActive from './icAddNewActive.png';
 import icLibrary from './icLibrary.png';
-import icLibraryActive from './icLibraryActive.png';
 import icMyRentals from './icMyRentals.png';
-import icMyRentalsActive from './icMyRentalsActive.png';
 import icSettings from './icSettings.png';
-import icSettingsActive from './icSettingsActive.png';
 import icWishlist from './icWishlist.png';
-import icWishlistActive from './icWishlistActive.png';
 
 const iconTabBar = {
   icAddNew,
-  icAddNewActive,
   icLibrary,
-  icLibraryActive,
   icMyRentals,
-  icMyRentalsActive,
   icSettings,
-  icSettingsActive,
-  icWishlist,
-  icWishlistActive
+  icWishlist
 };
 
 export default iconTabBar;
