+++ wbooks/src/app/components/customTabBarIcon/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+import { cerulean, dustyGray } from '@constants/colors';
+
+export default StyleSheet.create({
+  active: {
+    tintColor: cerulean
+  },
+  inactive: {
+    tintColor: dustyGray
+  }
+});
